+++ internal/resolvers/twitter/resolver.go
@@ -9,6 +9,7 @@ import (
 	"github.com/Chatterino/api/pkg/cache"
 	"github.com/Chatterino/api/pkg/config"
 	"github.com/Chatterino/api/pkg/resolver"
+	"github.com/Chatterino/api/pkg/utils"
 )
 
 const (
@@ -34,6 +35,25 @@ const (
 var (
 	tweetRegexp       = regexp.MustCompile(`(?i)\/.*\/status(?:es)?\/([^\/\?]+)`)
 	twitterUserRegexp = regexp.MustCompile(`(?i)twitter\.com\/([^\/\?\s]+)(\/?$|(\?).*)`)
+
+	/* These routes refer to non-user pages. If the capture group in twitterUserRegexp
+	   matches any of these names, we must not resolve it as a Twitter user link.
+
+	   The pages are listed alphabetically. They were sourced by simply looking around the
+	   Twitter web page. AFAIK, there is no resource describing these "special" routes.
+	*/
+	nonUserPages = utils.SetFromSlice([]interface{}{
+		"compose",
+		"explore",
+		"home",
+		"logout",
+		"messages",
+		"notifications",
+		"search",
+		"settings",
+		"tos",
+		"privacy",
+	})
 
 	tweetTooltipTemplate = template.Must(template.New("tweetTooltip").Parse(tweetTooltip))